# docs/openapi.yaml
components:
  schemas:
...
    Decorator:
      description: "Represents a torero decorator.\n\nDecorators modify the behavior\
        \ of torero services by adding functionality\nsuch as authentication, logging,\
        \ or parameter validation.\n\nAttributes:\n    name: Unique identifier for\
        \ the decorator\n    description: Human-readable explanation of the decorator's\
        \ purpose\n    type: Category of decorator indicating its function\n    parameters:\
        \ Schema for parameters accepted by the decorator\n    registries: Optional\
        \ metadata about where the decorator is registered"
      example:
        description: Adds HTTP Basic Authentication to service calls
        name: auth-basic
        parameters:
          password:
            required: true
            secret: true
            type: string
          username:
            required: true
            type: string
        registries:
          file:
            path: /etc/torero/decorators/auth-basic
        type: authentication
      properties:
        description:
          anyOf:
          - type: string
          - type: 'null'
          description: Human-readable explanation of the decorator's purpose
          title: Description
        name:
          description: Unique identifier for the decorator
          title: Name
          type: string
        parameters:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          description: Schema for parameters accepted by the decorator
          title: Parameters
        registries:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          description: Metadata about where the decorator is registered
          title: Registries
        type:
          description: Category of decorator indicating its function
          title: Type
          type: string
      required:
      - name
      - type
      title: Decorator
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    Repository:
      description: "Represents a torero repository.\n\nRepositories are storage locations\
        \ for torero components such as\nservices, decorators, and other resources.\n\
        \nAttributes:\n    name: Unique identifier for the repository\n    description:\
        \ Human-readable explanation of the repository's purpose\n    type: Category\
        \ of repository (e.g., 'file', 'git', 's3')\n    location: Location URI for\
        \ the repository\n    metadata: Optional additional metadata about the repository"
      example:
        description: Local file repository for services
        location: /etc/torero/services
        metadata:
          created: '2023-01-01T00:00:00Z'
          owner: torero
        name: local-services
        type: file
      properties:
        description:
          anyOf:
          - type: string
          - type: 'null'
          description: Human-readable explanation of the repository's purpose
          title: Description
        location:
          description: Location URI for the repository
          title: Location
          type: string
        metadata:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          description: Additional metadata about the repository
          title: Metadata
        name:
          description: Unique identifier for the repository
          title: Name
          type: string
        type:
          description: Category of repository (e.g., 'file', 'git', 's3')
          title: Type
          type: string
      required:
      - name
      - type
      - location
      title: Repository
      type: object
    Secret:
      description: "Represents a torero secret.\n\nSecrets are sensitive values stored\
        \ securely and used by torero\nservices and decorators.\n\nAttributes:\n \
        \   name: Unique identifier for the secret\n    description: Human-readable\
        \ explanation of the secret's purpose\n    type: Category of secret (e.g.,\
        \ 'password', 'api-key', 'token')\n    created_at: Timestamp when the secret\
        \ was created\n    metadata: Optional additional metadata about the secret"
      example:
        created_at: '2023-01-01T00:00:00Z'
        description: Database password for the application
        metadata:
          owner: admin
          provider: vault
        name: db-password
        type: password
      properties:
        created_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          description: Timestamp when the secret was created
          title: Created At
        description:
          anyOf:
          - type: string
          - type: 'null'
          description: Human-readable explanation of the secret's purpose
          title: Description
        metadata:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          description: Additional metadata about the secret
          title: Metadata
        name:
          description: Unique identifier for the secret
          title: Name
          type: string
        type:
          description: Category of secret (e.g., 'password', 'api-key', 'token')
          title: Type
          type: string
      required:
      - name
      - type
      title: Secret
      type: object
    Service:
      description: "Represents a torero service.\n\nA service is a runnable unit in\
        \ torero that can be executed to perform specific tasks.\nServices have different\
        \ types (e.g., ansible-playbook, opentofu-plan, python-script)\nand can be\
        \ tagged for organization and filtering.\n\nAttributes:\n    name: Unique\
        \ identifier for the service\n    description: Human-readable explanation\
        \ of the service's purpose\n    type: Category of service indicating the underlying\
        \ technology\n    tags: List of labels for grouping and filtering services\n\
        \    registries: Optional metadata about where the service is registered"
      example:
        description: Backs up configurations from network devices
        name: network-backup
        registries:
          file:
            path: /etc/torero/services/network-backup
        tags:
        - network
        - backup
        - daily
        type: ansible-playbook
      properties:
        description:
          anyOf:
          - type: string
          - type: 'null'
          description: Human-readable explanation of the service's purpose
          title: Description
        name:
          description: Unique identifier for the service
          title: Name
          type: string
        registries:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          description: Metadata about where the service is registered
          title: Registries
        tags:
          description: List of labels for grouping and filtering services
          items:
            type: string
          title: Tags
          type: array
        type:
          description: Category of service indicating the underlying technology
          title: Type
          type: string
      required:
      - name
      - type
      - tags
      title: Service
      type: object
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  contact:
    email: opensource@itential.com
    name: torero Development Team
    url: https://torero.dev/contact
  description: "\n        RESTful API for interacting with torero services.\n    \
    \    \n        This API provides endpoints for discovering and managing torero\
    \ services,\n        allowing you to list services, filter by type or tag, and\
    \ get metadata\n        about available service types and tags.\n        \n  \
    \      Model Context Protocol (MCP) Integration:\n        This API follows OpenAPI\
    \ standards and provides comprehensive type\n        information for seamless\
    \ integration with MCP-enabled applications.\n        "
  license:
    name: Apache License 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  title: torero API
  version: 0.1.0
  x-mcp-compatible: true
  x-mcp-description: This API is optimized for Model Context Protocol integration.
  x-mcp-version: '1.0'
openapi: 3.1.0
paths:
...
  /v1/decorators/:
    get:
      description: "Get all registered torero decorators with optional filtering.\n\
        \    \n    This endpoint returns a list of all decorators registered with\
        \ torero.\n    You can filter the results by decorator type to narrow down\
        \ the list.\n    \n    Examples:\n    - List all decorators: GET /v1/decorators/\n\
        \    - List all authentication decorators: GET /v1/decorators/?type=authentication"
      operationId: list_decorators_v1_decorators__get
      parameters:
      - description: Filter by decorator type, e.g. 'authentication'
        in: query
        name: type
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by decorator type, e.g. 'authentication'
          title: Type
      - description: Number of items to skip
        in: query
        name: skip
        required: false
        schema:
          default: 0
          description: Number of items to skip
          minimum: 0
          title: Skip
          type: integer
      - description: Maximum number of items to return
        in: query
        name: limit
        required: false
        schema:
          default: 100
          description: Maximum number of items to return
          maximum: 1000
          minimum: 1
          title: Limit
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Decorator'
                title: Response List Decorators V1 Decorators  Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List decorators
      tags:
      - decorators
...
  /v1/repositories/:
    get:
      description: "Get all registered torero repositories with optional filtering.\n\
        \    \n    This endpoint returns a list of all repositories registered with\
        \ torero.\n    You can filter the results by repository type to narrow down\
        \ the list.\n    \n    Examples:\n    - List all repositories: GET /v1/repositories/\n\
        \    - List all file repositories: GET /v1/repositories/?type=file\n    -\
        \ List all git repositories: GET /v1/repositories/?type=git"
      operationId: list_repositories_v1_repositories__get
      parameters:
      - description: Filter by repository type, e.g. 'file', 'git', 's3'
        in: query
        name: type
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by repository type, e.g. 'file', 'git', 's3'
          title: Type
      - description: Number of items to skip
        in: query
        name: skip
        required: false
        schema:
          default: 0
          description: Number of items to skip
          minimum: 0
          title: Skip
          type: integer
      - description: Maximum number of items to return
        in: query
        name: limit
        required: false
        schema:
          default: 100
          description: Maximum number of items to return
          maximum: 1000
          minimum: 1
          title: Limit
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Repository'
                title: Response List Repositories V1 Repositories  Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List repositories
      tags:
      - repositories
...
  /v1/secrets/:
    get:
      description: "Get all registered torero secrets with optional filtering.\n \
        \   \n    This endpoint returns a list of all secrets registered with torero.\n\
        \    You can filter the results by secret type to narrow down the list.\n\
        \    \n    Examples:\n    - List all secrets: GET /v1/secrets/\n    - List\
        \ all password secrets: GET /v1/secrets/?type=password\n    - List all API\
        \ key secrets: GET /v1/secrets/?type=api-key"
      operationId: list_secrets_v1_secrets__get
      parameters:
      - description: Filter by secret type, e.g. 'password', 'api-key', 'token'
        in: query
        name: type
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by secret type, e.g. 'password', 'api-key', 'token'
          title: Type
      - description: Number of items to skip
        in: query
        name: skip
        required: false
        schema:
          default: 0
          description: Number of items to skip
          minimum: 0
          title: Skip
          type: integer
      - description: Maximum number of items to return
        in: query
        name: limit
        required: false
        schema:
          default: 100
          description: Maximum number of items to return
          maximum: 1000
          minimum: 1
          title: Limit
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Secret'
                title: Response List Secrets V1 Secrets  Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List secrets
      tags:
      - secrets
...
  /v1/services/:
    get:
      description: "Get all registered torero services with optional filtering.\n\
        \    \n    This endpoint returns a list of all services registered with torero.\n\
        \    You can filter the results by service type and/or tag to narrow down\n\
        \    the list to specific services of interest.\n    \n    Examples:\n   \
        \ - List all services: GET /v1/services/\n    - List all ansible playbooks:\
        \ GET /v1/services/?type=ansible-playbook\n    - List services with \"network\"\
        \ tag: GET /v1/services/?tag=network\n    - Combine filters: GET /v1/services/?type=ansible-playbook&tag=network"
      operationId: list_services_v1_services__get
      parameters:
      - description: Filter by service type, e.g. 'ansible-playbook'
        in: query
        name: type
        required: false
        schema:
          anyOf:
          - enum:
            - ansible-playbook
            - opentofu-plan
            - python-script
            type: string
          - type: 'null'
          description: Filter by service type, e.g. 'ansible-playbook'
          examples:
            ansible-playbook:
              description: Filter for only Ansible Playbook services
              summary: Ansible Playbook
              value: ansible-playbook
            opentofu-plan:
              description: Filter for only OpenTofu Plan services
              summary: OpenTofu Plan
              value: opentofu-plan
            python-script:
              description: Filter for only Python Script services
              summary: Python Script
              value: python-script
          title: Type
      - description: Filter by tag (e.g., 'network', 'backup')
        in: query
        name: tag
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by tag (e.g., 'network', 'backup')
          examples:
            automation:
              description: Filter services with the 'automation' tag
              summary: Automation Tag
              value: automation
            backup:
              description: Filter services with the 'backup' tag
              summary: Backup Tag
              value: backup
            network:
              description: Filter services with the 'network' tag
              summary: Network Tag
              value: network
          title: Tag
      - description: Number of items to skip
        in: query
        name: skip
        required: false
        schema:
          default: 0
          description: Number of items to skip
          minimum: 0
          title: Skip
          type: integer
      - description: Maximum number of items to return
        in: query
        name: limit
        required: false
        schema:
          default: 100
          description: Maximum number of items to return
          maximum: 1000
          minimum: 1
          title: Limit
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/Service'
                title: Response List Services V1 Services  Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List services
      tags:
      - services
...